# api/openapi/marketplace.v1.yaml
openapi: 3.1.0
info:
  title: Marketplace API
  version: 1.0.0
  description: Публичный REST API маркетплейса. Контракт является источником правды (contract-first).
servers:
  - url: https://api.marketplace.kz
  - url: http://localhost:8080
...
paths:
...
  /api/v1/cart:
    get:
      tags: [cart]
      operationId: getCart
      description: "Корзина авторизованного пользователя или устройства из X-Device-ID с актуальными ценами, наличием и доставкой по продавцам."
      security: [{}, { bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/DeviceID"
      responses:
        "200": { $ref: "#/components/responses/Cart" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/cart/items:
    post:
      tags: [cart]
      operationId: addCartItem
      description: Добавляет оффер или увеличивает количество позиции.
      security: [{}, { bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/DeviceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sku, quantity]
              properties:
                sku: { type: string, description: ID оффера }
                quantity: { type: integer, minimum: 1, maximum: 999 }
      responses:
        "200": { $ref: "#/components/responses/Cart" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /api/v1/cart/items/{sku}:
    patch:
      tags: [cart]
      operationId: updateCartItem
      description: Меняет количество позиции; 0 удаляет позицию.
      security: [{}, { bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/DeviceID"
        - $ref: "#/components/parameters/SKUPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity: { type: integer, minimum: 0, maximum: 999 }
      responses:
        "200": { $ref: "#/components/responses/Cart" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    delete:
      tags: [cart]
      operationId: removeCartItem
      security: [{}, { bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/DeviceID"
        - $ref: "#/components/parameters/SKUPath"
      responses:
        "200": { $ref: "#/components/responses/Cart" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/cart/promo-code:
    post:
      tags: [cart]
      operationId: applyCartPromoCode
      description: Сохраняет промокод, если котировка корзины его принимает.
      security: [{}, { bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/DeviceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
      responses:
        "200": { $ref: "#/components/responses/Cart" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    delete:
      tags: [cart]
      operationId: removeCartPromoCode
      security: [{}, { bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/DeviceID"
      responses:
        "200": { $ref: "#/components/responses/Cart" }
  /api/v1/cart/checkout-preview:
    post:
      tags: [cart]
      operationId: previewCheckout
      description: Корзина с расчётом доставки выбранным способом и признаком готовности к оформлению.
      security: [{}, { bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/DeviceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                delivery_method: { type: string, enum: [standard] }
      responses:
        "200": { $ref: "#/components/responses/Cart" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/cart/merge:
    post:
      tags: [cart]
      operationId: mergeCart
      description: Переносит анонимную корзину устройства в корзину пользователя после входа.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/DeviceID"
      responses:
        "200": { $ref: "#/components/responses/Cart" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DeviceID:
      name: X-Device-ID
      in: header
      description: Идентификатор устройства для анонимной корзины
      schema: { type: string, pattern: "^[A-Za-z0-9_-]{16,128}$" }
    SKUPath:
      name: sku
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Cart:
      description: Корзина
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Cart" }
...
    BadRequest:
      description: Некорректный запрос
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Требуется аутентификация
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    NotFound:
      description: Ресурс не найден
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    UnprocessableEntity:
      description: Нарушено бизнес-правило
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        instance: { type: string }
        trace_id: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, code, message]
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
...
    Cart:
      type: object
      required: [delivery_method, groups, items_count, subtotal, discount, shipping, total, issues, ready]
      properties:
        cart_id: { type: string, format: uuid }
        currency: { type: string }
        promo_code: { type: string }
        promo_error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
        delivery_method: { type: string }
        groups:
          type: array
          items:
            type: object
            required: [seller_id, items, subtotal, discount, shipping, total]
            properties:
              seller_id: { type: string, format: uuid }
              subtotal: { type: integer, format: int64 }
              discount: { type: integer, format: int64 }
              shipping: { type: integer, format: int64 }
              total: { type: integer, format: int64 }
              items:
                type: array
                items:
                  type: object
                  required: [sku, quantity, available, status, unit_price, saved_price, price_changed, subtotal, discount, total]
                  properties:
                    sku: { type: string }
                    product_id: { type: string }
                    title: { type: string }
                    cover_key: { type: string }
                    quantity: { type: integer }
                    available: { type: integer }
                    status: { type: string, enum: [available, insufficient_stock, unavailable] }
                    unit_price: { type: integer, format: int64 }
                    saved_price: { type: integer, format: int64 }
                    price_changed: { type: boolean }
                    compare_at: { type: integer, format: int64 }
                    subtotal: { type: integer, format: int64 }
                    discount: { type: integer, format: int64 }
                    total: { type: integer, format: int64 }
        items_count: { type: integer }
        subtotal: { type: integer, format: int64 }
        discount: { type: integer, format: int64 }
        shipping: { type: integer, format: int64 }
        total: { type: integer, format: int64 }
        issues:
          type: array
          items:
            type: object
            required: [sku, code]
            properties:
              sku: { type: string }
              code: { type: string, enum: [CART_OFFER_UNAVAILABLE, CART_QUANTITY_EXCEEDS_STOCK] }
        ready: { type: boolean }
        updated_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
...